Refactor GLSL generation spec.
